Add basic tuplets
Only working so far when the tuplet is the only inhabitant of the bar.
Attempted with auto-beaming which seems OK for triplets in 3/4 but not
quite right for quadruplets in 6/8.

Need to change the map over the music contents of a bar to a fold so
that we can pass in the position of the tuplet (from the notes already
translated in the bar).