Add TickableContext
VexFlow has a concept of a 'tickable' which is any item in a bar stave
which has a note-like identity and takes up space within the stave
(essentially just notes or rests).  Add a Type 'TickableContext' which
represents both the count of such items and their duration for any ABC
music item which is tickable.  Add this to the MusicSpec type and
accumulate it when traversing a bar of 'musics'.  This means replacing
the map with a left fold.

We really need to give MusicSpec itself a Monoid instance to simplify
this.