Start threading state through the translation
New module TranslateStateful which uses the type:

type Translation a = ExceptT String (State AbcContext) a

This allows the (changing) ABC context to be threaded through the computation.  So far, just exposed translating a single bar to prove the basic concept.
I'll keep the non-stateful equivalents for the time being so we can
check we're not losing anything.