Prepare for variable length bars
The accumulated bar lengths are saved into AbcContext and thus threaded
through the translation. Added width and xOffset to BarSpec, although at
the moment, the width is hard-coded to 250 for each bar.  Use this value
when configuraing the bar stave.

We need to improve Score.purs such that AbcContext is never a parameter
to any of the FFI methods.  This is because it does not use the threaded
value.  Instead, track any required values in the translation result.