Fix bug estimating the bar width
Bars which were essentially empty on the first stave were being given a
width, although the width should be zero.  This was because we tested
for the presence of a clef by looking at stave 0, whereas it should be
both stave 0 and bar 0.

Fixed by simplifying - the initial config sets 'isNewTimeSignature' true
at startoff (which is the case) and so we don't need any longer to check
the stave or bar number.

This context is set off after bar 0 and set back on if there is an
intermediate header which sets the new time signature.