Sort out volta that ends at the stave end
I remember now why I had the spurious code unilaterally to end a volta
if a bar was empty of contents.  This was because the first bar in a new
stave is often empty when it just displays the key signature.

Amended the code so that if the first bar in a stave is empty, then any
volta is suspended (not displayed). When the first bar with actual
contents is processed, the volta is continued or ended, as appropriate.

This effectively means that there will never be a volta line above such
an empty first bar.