Simplify the handling of VexFlow durations
VexFlow uses, on different occasions, a compound string for durations
(for example "qdd" for a doubly-dotted quarter note) or a simple string
"q" plus a dot count if 2.

This commit introduces a VexDuration type for the first such
representation, allowing you to generate the second from it when needed,
thus simplifying the code a little.