Replace auto-beaming with explicit beaming
This has the potential to give us the control we need over beaming which
tends to be lost when tuplets are involved.  The strategy is to define
'standard beams' and 'tuplet beams' separately and then combine them
into a single set of beam groups.

There are things to be wary of - for example more chance of raising a
runtime error in VexFlow.  Note that beams are prevented if any note is
longer than a quarter note.

The strategy for standard beams is to restrict them to exist within a
beat whereas tupleted beams can exist across beats if the notes are all
of the requisite length.