Make the stave line colour darker
At the moment, the lightish gray colour does not show up well on
printing.  It seems we can't tamper with the line width in VexFlow but
we can set the colour by means of a fill_style attribute in
staveOptions.  Accordingly, set a lineColour property in staveConfig
which allows this to be altered (but is as yet not exposed to users).