Revert to original VexFlow formatting
The default formatter setting is hopeless for our case, giving too much
real estate to minims and crotchets and too little to beamed eighth and
sixtenth notes.

This commit attempts to revert to the formatting style that we had in
VexFlow 1.2.89.  It needs to use a magic setting in softmaxFactor and we
can no longer use the default formatAndDraw static method which uses the
new VexFlow formatting introduced in version 3 (and 4).